+++ test/markdown-link-check.test.js
@@ -60,7 +60,7 @@ describe('markdown-link-check', function () {
     });
 
     it('should check the links in sample.md', function (done) {
-        markdownLinkCheck(fs.readFileSync(path.join(__dirname, 'sample.md')).toString().replace(/%%BASE_URL%%/g, baseUrl), { baseUrl: baseUrl, httpHeaders: [{ urls: [baseUrl + '/basic-auth'], headers: { 'Authorization': 'Basic Zm9vOmJhcg==', 'Foo': 'Bar' }}] }, function (err, results) {
+        markdownLinkCheck(fs.readFileSync(path.join(__dirname, 'sample.md')).toString().replace(/%%BASE_URL%%/g, baseUrl), { baseUrl: baseUrl, ignorePatterns: [{ pattern: /not-working-and-ignored/ }], httpHeaders: [{ urls: [baseUrl + '/basic-auth'], headers: { 'Authorization': 'Basic Zm9vOmJhcg==', 'Foo': 'Bar' }}] }, function (err, results) {
             expect(err).to.be(null);
             expect(results).to.be.an('array');
 
@@ -85,6 +85,9 @@ describe('markdown-link-check', function () {
 
                 // basic-auth
                 { statusCode: 200, status: 'alive' },
+
+                // ignored
+                { statusCode: 0, status: 'ignored' },
 
                 // hello image
                 { statusCode: 200, status: 'alive' },